+++ .github/workflows/feature_test.yml
@@ -56,6 +56,8 @@ jobs:
           git fetch --depth=1 --no-tags origin $GITHUB_BASE_REF:$GITHUB_BASE_REF
           git fetch --depth=1 --no-tags origin $GITHUB_HEAD_REF:$GITHUB_HEAD_REF
           pre-commit run --files $(git diff --diff-filter=d --name-only $GITHUB_BASE_REF..$GITHUB_HEAD_REF)
+        env:
+          SKIP: debug-statements
 
   unit_tests:
     name: "Unit tests: Python ${{ matrix.python-version }}"
